+++ NEWS.md
@@ -4,6 +4,7 @@ ctmm 1.3.1 (2026-05-06)
 * new home-range index() function
 * mean() now works when range=FALSE
 * pkde() can now take fixed weights
+* pkde() now has a finite population correction argument
 * agde() now works on multiple individuals
 * as.telemetry() can now import light-level geolocator error ellipses
 * as.telemetry() no longer requires strict Argos formatting